Fixed issue in post build where manifest is not deleted if there is an exception in post build execute.
Added functionality to transform view model annotations using regex, from [Display(Name = "Test1", Prompt="Test2", ShortName="Test3")] to [Display(_("Test1"), _("Test2"), _("Test3"))] so that the annotations can be processed by xgettext, after which point the annotations are restored to their original values.

+++ src/i18n/PostBuildTask.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 
 namespace i18n
 {
@@ -12,6 +13,7 @@ namespace i18n
     /// </summary>
     public class PostBuildTask
     {
+        private const string cTransformedFileExtension = "PostBuildBak";
         ///<summary>
         /// Runs GNU xgettext to extract a messages template file
         ///</summary>
@@ -23,13 +25,20 @@ namespace i18n
         {
             if (inputPaths == null || inputPaths.Length == 0) inputPaths = new string[] { outputPath };
 
-            var manifest = BuildProjectFileManifest(inputPaths);
+            string manifest = null;  
+            try
+            {
+                manifest = BuildProjectFileManifest(inputPaths);
 
             CreateMessageTemplate(outputPath, manifest, gettext);
 
             MergeTemplateWithExistingLocales(outputPath, msgmerge);
-
-            File.Delete(manifest);
+            }
+            finally
+            {
+                if(manifest !=  null) File.Delete(manifest);
+                RestoreTransformBackups(inputPaths);
+            }
         }
 
         private static void MergeTemplateWithExistingLocales(string outputPath, string options)
@@ -103,6 +112,7 @@ namespace i18n
         private static string BuildProjectFileManifest(string[] paths)
         {
             var cs = GetFiles(paths, "*.cs");
+            TransformAnnotations(cs);
             var razor = GetFiles(paths, "*.cshtml");
             var files = (new[] { cs, razor }).SelectMany(f => f).ToList();
 
@@ -115,6 +125,85 @@ namespace i18n
                 }
             }
             return temp;
+        }
+
+        private void RestoreTransformBackups(string[] paths)
+        {
+            var files = GetFiles(paths, "*.cs" + cTransformedFileExtension);
+            foreach (var file in files)
+            {
+                var originalFileName = file.Remove(file.Length - cTransformedFileExtension.Length);
+                File.Delete(originalFileName);
+                File.Move(file, originalFileName);
+            }
+        }
+
+        private static void TransformAnnotations(string[] paths)
+        {
+            foreach (var path in paths)
+            {
+                string fileText = File.ReadAllText(path, encoding: System.Text.Encoding.UTF8);
+                string transformed = null;
+                if (TransformAnnotations(fileText, out transformed))
+                {
+                    if (!File.Exists(path + cTransformedFileExtension))
+                    {
+                        File.WriteAllText(path: path + cTransformedFileExtension, contents: fileText, encoding: System.Text.Encoding.UTF8);
+                    }
+                    File.WriteAllText(path: path, contents: transformed, encoding: System.Text.Encoding.UTF8);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Transform annotations in code file to format than can be processed by xgettext
+        /// </summary>
+        /// <param name="fileText">original code file</param>
+        /// <param name="transformed">transformed code file</param>
+        /// <returns>true if input file contained any annotations that were transformed</returns>
+        private static bool TransformAnnotations(string fileText, out string transformed)
+        {
+            Regex displayName = new Regex(@"\[(i18n.DataAnnotations.)?Display\((?<Property>.*)Name\s?=\s?(?<String>\""[^\""]*\"")(?<Others>.*)");
+            Regex displayPrompt = new Regex(@"\[(i18n.DataAnnotations.)?Display\((?<Property>.*)Prompt\s?=\s?(?<String>\""[^\""]*\"")(?<Others>.*)");
+            Regex displayDescription = new Regex(@"\[(i18n.DataAnnotations.)?Display\((?<Property>.*)Description\s?=\s?(?<String>\""[^\""]*\"")(?<Others>.*)");
+            Regex dataType = new Regex(@"\[(i18n.DataAnnotations.)?DataType\((?<Property>.*)ErrorMessage\s?=\s?(?<String>\""[^\""]*\"")(?<Others>.*)\)");
+            Regex required = new Regex(@"\[(i18n.DataAnnotations.)?Required\((?<Property>.*)ErrorMessage\s?=\s?(?<String>\""[^\""]*\"")(?<Others>.*)\)");
+            Regex stringLength = new Regex(@"\[(i18n.DataAnnotations.)?StringLength\((?<Property>.*)ErrorMessage\s?=\s?(?<String>\""[^\""]*\"")(?<Others>.*)\)");
+            transformed = fileText;
+            bool localizeAnnotations = false;
+
+            if (displayName.IsMatch(transformed))
+            {
+                transformed = displayName.Replace(input: transformed, replacement: @"[Display(${Property}_(${String})${Others}");
+                localizeAnnotations = true;
+            }
+            if (displayPrompt.IsMatch(transformed))
+            {
+                transformed = displayPrompt.Replace(input: transformed, replacement: @"[Display(${Property}_(${String})${Others}");
+                localizeAnnotations = true;
+            }
+            if (displayDescription.IsMatch(transformed))
+            {
+                transformed = displayDescription.Replace(input: transformed, replacement: @"[Display(${Property}_(${String})${Others}");
+                localizeAnnotations = true;
+            }
+            if (dataType.IsMatch(transformed))
+            {
+                transformed = dataType.Replace(input: transformed, replacement: @"[DataType(${Property}_(${String})${Others})");
+                localizeAnnotations = true;
+            }
+            if (required.IsMatch(transformed))
+            {
+                transformed = required.Replace(input: transformed, replacement: @"[Required(${Property}_(${String})${Others})");
+                localizeAnnotations = true;
+            }
+            if (stringLength.IsMatch(transformed))
+            {
+                transformed = stringLength.Replace(input: transformed, replacement: @"[StringLength(${Property}_(${String})${Others})");
+                localizeAnnotations = true;
+            }
+
+            return localizeAnnotations;
         }
     }
 }
